requirements.txt
alembic==1.11.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.1.0
cymem==2.0.7
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.1.2
executing==1.2.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
ipython==8.14.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
langcodes==3.3.0
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
murmurhash==1.0.9
mysql-connector-python==8.0.33
numpy==1.25.0
outcome==1.2.0
packaging==23.1
parso==0.8.3
pathy==0.10.2
pickleshare==0.7.5
Pillow==10.0.0
pluggy==1.2.0
preshed==3.0.8
prettytable==3.8.0
prompt-toolkit==3.0.38
protobuf==3.20.3
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.10
Pygments==2.15.1
PyMySQL==1.1.0
PySocks==1.7.1
pytest==7.4.0
requests==2.31.0
selenium==4.10.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.17
sqlparse==0.4.4
srsly==2.4.6
stack-data==0.6.2
thinc==8.1.10
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.1
trio-websocket==0.10.3
typer==0.9.0
typing_extensions==4.7.0
urllib3==2.0.3
wasabi==1.1.2
wcwidth==0.2.6
Werkzeug==2.3.6
wsproto==1.2.0
WTForms==3.0.1
